Initial Commit of practice file
This is the collection of exercises that I have completed from the
Second Edition of The ANSI C Programming Langugae book by Brian W. Kernighan
and Dennis M. Ritchie.

As a way to sharpen my C skills I have dedicated myself to completing all of
the coding exercises that are contained in this book. I will be updating
it periodically as I complete more exercises and refine the programs from old
exercises